requirements.txt
asgiref==3.8.1
Django
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
psycopg
psycopg2-binary
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
PyJWT==1.7.1
environs==9.5.0